+++ src/Services/Facebook.php
@@ -17,12 +17,13 @@ class Facebook extends AbstractService
 {
 
     /** @var string */
-    protected $base = 'https://www.facebook.com/sharer/sharer.php?u=';
+    protected $base;
     /** @var string */
     protected $icon;
 
     public function __construct()
     {
+        $this->base = Config::get('social-media-links.services.facebook.base_url');
         $this->icon = Config::get('social-media-links.services.facebook.icon');
     }